requirements.txt
Flask==1.1.4
python-dotenv
psycopg2-binary
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
pylint
pylint-flask
pylint_flask_sqlalchemy
autopep8